requirements.txt
# PIP Requirements (for a development/build/test environment, Python 2.7).

Django==1.8.9
djangorestframework==3.3.2
drf-extensions==0.2.8
